Adding in some iteam to loop throught

+++ src/main/java/com/pluralsight/Product.java
@@ -3,9 +3,9 @@ package com.pluralsight;
 public class Product {
     private int id;
     private String name;
-    private float price;
+    private double price;
 
-    public Product(int id, String name, float price) {
+    public Product(int id, String name, double price) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -27,11 +27,11 @@ public class Product {
         this.name = name;
     }
 
-    public float getPrice() {
+    public double getPrice() {
         return price;
     }
 
-    public void setPrice(float price) {
+    public void setPrice(double price) {
         this.price = price;
     }
 }
